perf: optimize firmware size, RAM usage, and rendering performance
- Reduce Flash size by ~27.2 KB and RAM by 1.5 KB
- Add horizontal run-length acceleration to Thai font rendering
- Unify Thai font renderer to eliminate duplicate parsing code
- Optimize BearSSL TLS buffer size to 1024/512 for Gold API
- Prevent clock flicker in updateTimeOnly
- Minify embedded INDEX_HTML
- Optimize BitmapGlyphIndex struct size by 50%
- Add F() and PSTR() to all static strings
- Prevent heap fragmentation in buildPromptPayPayload

+++ smart_clock_esp8266/ThaiFont.h
@@ -20,7 +20,7 @@
 
 #define THAIFONT_TONES_HEIGHT 3
 
-typedef struct { uint32_t codepoint; uint32_t offset; } BitmapGlyphIndex;
+typedef struct { uint16_t codepoint; uint16_t offset; } BitmapGlyphIndex;
 
 const uint8_t ThaiFontBitmap[] PROGMEM = {
   0x00, 0x00, 0x00, 0x00, 0xC0, 0xC0, 0xC0, 0xC0, 0xC0, 0xC0, 0xC0, 0x00, 0xC0, 0xC0, 0x00, 0x00,  // U+0021 "!"
